Added grid alignment system

+++ src/ecs/Components/Position.java
@@ -2,8 +2,8 @@ package ecs.Components;
 
 public class Position extends Component{
 
-    public float i;
-    public float j;
+    public int i;
+    public int j;
 
     public Position(int i, int j){
         this.i = i;
